Add fuzziness parsing utilities and GNOME Shell version parser
- Implemented `fuzzinessFromEnumIndex` to convert enum index to Fuzziness.
- Added `parseFuzziness` function to handle string and Fuzziness inputs.
- Created `parseGnomeShellVersionString` to extract major version from GNOME Shell version strings.
