+++ src/constants/times/extension.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import { extensionGettext } from "../../utils/gettext";
+import { extensionGettext } from "../../utils/gettext/gettext_utils_ext.js";
 import { createTimeConstants } from "../times/core.js";
 
 const fns = extensionGettext;
