+++ src/constants/times/prefs.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import { prefsGettext } from "../../utils/gettext";
+import { prefsGettext } from "../../utils/gettext/gettext_utils_prefs.js";
 import { createTimeConstants } from "../times/core.js";
 
 const fns = prefsGettext;
