+++ src/prefs/pages/general_page/formatters.ts
@@ -1,10 +1,10 @@
 import Gtk from "gi://Gtk";
 import { ClockFormatter, TimeFormat } from "../../../core/clock_formatter.js";
-import { fuzzinessFromEnumIndex } from "../../../utils/parse/index.js";
+import { fuzzinessFromEnumIndex } from "../../../utils/parse_utils.js";
 import SettingsKey from "../../../models/settings_keys.js";
 import { TRANSLATE_PACK } from "../../helpers.js";
 import { addComboRow } from "../../ui/rows.js";
-import { prefsGettext } from "../../../utils/gettext/index.js";
+import { prefsGettext } from "../../../utils/gettext/gettext_utils_prefs.js";
 
 export function getTimeFormatsList(settings: any): Gtk.StringList {
   /**
