+++ src/services/settings_manager.ts
@@ -15,7 +15,7 @@ import GObject from "gi://GObject";
 import { SETTINGS } from "../constants/index.js";
 import SettingsKey from "../models/settings_keys.js";
 import { logWarn, logErr } from "../utils/error_utils.js";
-import { fuzzinessFromEnumIndex } from "../utils/parse/index.js";
+import { fuzzinessFromEnumIndex } from "../utils/parse_utils.js";
 import { Fuzziness } from "../core/clock_formatter.js";
 
 /**
