+++ src/services/style_service.ts
@@ -11,7 +11,8 @@
  */
 
 import Gio from "gi://Gio";
-import { normalizeColor, accentNameToHex } from "../utils/color/index.js";
+import { normalizeColor } from "../utils/color/color_utils.js";
+import { accentNameToHex } from "../utils/color/accent_color_utils.js";
 import { logInfo, logWarn } from "../utils/error_utils.js";
 import { getDividerText } from "../constants/index.js";
 import SettingsKey from "../models/settings_keys.js";
