+++ src/tests/unit/color_utils.test.ts
@@ -1,4 +1,4 @@
-import { normalizeColor, escapeMarkup } from "../../utils/color";
+import { normalizeColor, escapeMarkup } from "../../utils/color/color_utils.js";
 
 describe("normalizeColor", () => {
   it("returns #ffffff for empty input", () => {
