+++ src/tests/unit/constants/times/core.test.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { createTimeConstants } from "../../../../constants/times/core.js";
-import type { GettextFunctions } from "../../../../utils/gettext";
+import type { GettextFunctions } from "../../../../utils/gettext/gettext_utils.js";
 
 describe("createTimeConstants", () => {
   let mockGettextFunctions: GettextFunctions;
