+++ src/tests/unit/fuzziness_utils.test.ts
@@ -1,7 +1,7 @@
 import {
   fuzzinessFromEnumIndex,
   parseFuzziness,
-} from "../../utils/parse/index.js";
+} from "../../utils/parse_utils.js";
 import { Fuzziness } from "../../core/clock_formatter.js";
 
 // Mock FUZZINESS_ENUM_MINUTES for testing
