--- src/utils/color/index.ts
@@ -1,2 +1,0 @@
-export * from "./color_utils.js";
-export * from "./accent_color_utils.js";
