--- src/utils/gettext/index.ts
@@ -1,3 +1,0 @@
-export * from "./gettext_utils.js";
-export * from "./gettext_utils_ext.js";
-export * from "./gettext_utils_prefs.js";
