--- src/utils/logging/index.ts
@@ -1,3 +1,0 @@
-export * from "./logger_interface.js";
-export * from "./logger_gjs.js";
-export * from "./logger_console.js";
