--- src/utils/parse/index.ts
@@ -1,2 +1,0 @@
-export * from "./fuzziness_utils.js";
-export * from "./shell_version_utils.js";
