--- src/utils/translate/index.ts
@@ -1,1 +1,0 @@
-export * from "./translate_pack_utils.js";
